+++ www/api/libs/utils/__init__.py
@@ -1,5 +1,4 @@
 from .abort import abort
-from .authorize import authorize
 from .api_response import api_response
 from .req_filters import process_filters
 from .sqlmodel_utils import (
