--- www/api/libs/utils/authorize/__init__.py
@@ -1,1 +1,0 @@
-from .authorize import authorize
