+++ www/api/modules/payments/infraestructure/controllers/subscription/subscription_http_controller.py
@@ -2,7 +2,8 @@ import json
 
 from flask import Blueprint, request
 
-from api.libs.utils import api_response as Response, as_dict, authorize
+from api.libs.utils import api_response as Response, as_dict
+from shared.utils.authorizer import authorizer
 
 from .... import application
 
@@ -10,7 +11,7 @@ subscription_api = Blueprint('subscription_api', __name__)
 
 
 @subscription_api.route('', methods=['POST'])
-@authorize()
+# @authorizer()
 def subscribe(user: dict):
     request_data = request.get_json()
