--- www/shared/__init__.py
@@ -1,1 +1,0 @@
-from .tipsy_jwt import verify_token
