+++ www/shared/domain/entity/__init__.py
@@ -1,0 +1,1 @@
+from .user_type import UserType
