+++ www/shared/domain/entity/user_type.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class UserType(Enum):
+    STUDENT = 'STUDENT'
+    TEACHER = 'TEACHER'
+    ADMIN = 'ADMIN'
