+++ www/webapp/static/css/layout/base.css
@@ -1,0 +1,14 @@
+html, body {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+.content {
+    flex: 1 1 auto;
+    overflow: scroll;
+}
+
+.h-full {
+    height: 100%;
+}
