+++ www/webapp/templates/layout/base.html
@@ -10,8 +10,9 @@
         <link 
             href="{{ url_for('static', filename='vendor/bootstrap/css/bootstrap.min.css') }}" 
             rel="stylesheet">
+        <link href="/static/css/layout/base.css" rel="stylesheet">
     </head>
-    <body data-bs-theme="light" id="body">
+    <body data-bs-theme="light" id="body" >
         {{ navbar() }}
 
         {% block body %}{% endblock %}
